+++ src/imports.ts
@@ -12,31 +12,8 @@ export const nitroImports: Preset[] = [
       'useStorage',
       'useNitroApp',
       'defineNitroPlugin',
-      'nitroPlugin'
-    ]
-  },
-  {
-    from: 'h3',
-    imports: [
-      'defineEventHandler',
-      'defineLazyEventHandler',
-      'eventHandler',
-      'lazyEventHandler',
-      'dynamicEventHandler',
-      'appendHeader',
-      'assertMethod',
-      'createError',
-      'handleCacheHeaders',
-      'isMethod',
-      'sendRedirect',
-      'useCookies',
-      'useCookie',
-      'deleteCookie',
-      'setCookie',
-      'useBody',
-      'useMethod',
-      'useQuery',
-      'useRawBody'
+      'nitroPlugin',
+      'defineRenderHandler'
     ]
   }
 ]