+++ test/fixture/nitro.config.ts
@@ -1,7 +1,8 @@
 import { defineNitroConfig } from '../../src'
 
 export default defineNitroConfig({
-  autoImport: {
+  compressPublicAssets: true,
+  imports: {
     presets: [
       {
         // TODO: move this to built-in preset
@@ -22,8 +23,12 @@ export default defineNitroConfig({
   ],
   prerender: {
     crawlLinks: true,
+    ignore: [
+      // '/api/param/'
+    ],
     routes: [
       '/prerender',
+      '/icon.png',
       '/404'
     ]
   }